Rename distillation namespace (#38391)
* Rename distillation function

* Rename namespace

* Update changelog
