+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
   - Update TLS version from 1.0 to 1.2
-  - Added support for Distillation jobs. Can be created by importing `disillation` from `azure.ai.ml.distillation`
+  - Added support for Distillation jobs. Can be created by importing `disillation` from `azure.ai.ml.model_customization`
 ### Bugs Fixed
   - Fix send email notification issue in model monitoring
