+++ sdk/ml/azure-ai-ml/tests/distillation_job/unittests/test_distillation_factory.py
@@ -2,7 +2,7 @@ import pytest
 
 from azure.ai.ml import Input
 from azure.ai.ml.constants import AssetTypes, DataGenerationTaskType, DataGenerationType
-from azure.ai.ml.distillation import distillation
+from azure.ai.ml.model_customization import distillation
 from azure.ai.ml.entities import NoneCredentialConfiguration, ServerlessConnection, WorkspaceConnection
 from azure.ai.ml.entities._job.distillation.distillation_job import DistillationJob
 from azure.ai.ml.entities._job.distillation.endpoint_request_settings import EndpointRequestSettings
